+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Recommended extensions for general use of languages to enhance the motive of recommending them
 - Option to view example debug configurations
 - Notes for various languages to convey potentially important information
+- Delete testcase button for running status
 
 ### Changed